requirements.txt
Pillow
tqdm
ConfigParser
